exponent in 64 bits

+++ include/core/Proof.hpp
@@ -8,11 +8,11 @@ namespace core {
 
 class Proof {
 public:
-    const uint64_t E;
-    const std::vector<uint64_t> B;
-    const std::vector<std::vector<uint64_t>> middles;
+    const uint32_t E;
+    const std::vector<uint32_t> B;
+    const std::vector<std::vector<uint32_t>> middles;
 
-    Proof(uint64_t exponent, std::vector<uint64_t> finalResidue, std::vector<std::vector<uint64_t>> intermediateResidues)
+    Proof(uint32_t exponent, std::vector<uint32_t> finalResidue, std::vector<std::vector<uint32_t>> intermediateResidues)
         : E(exponent), B(std::move(finalResidue)), middles(std::move(intermediateResidues)) {}
 
     // File I/O methods for proof files
@@ -20,11 +20,11 @@ public:
     static Proof load(const std::filesystem::path& filePath);
     
     // Hash functions for proof generation
-    static std::array<uint64_t, 4> hashWords(uint64_t E, const std::vector<uint64_t>& words);
-    static std::array<uint64_t, 4> hashWords(uint64_t E, 
+    static std::array<uint64_t, 4> hashWords(uint32_t E, const std::vector<uint32_t>& words);
+    static std::array<uint64_t, 4> hashWords(uint32_t E, 
                                            const std::array<uint64_t, 4>& hash,
-                                           const std::vector<uint64_t>& words);
-    static uint64_t res64(const std::vector<uint64_t>& words);
+                                           const std::vector<uint32_t>& words);
+    static uint64_t res64(const std::vector<uint32_t>& words);
 };
 
 } // namespace core
